+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * extend **Models**: add methods `read(URL)`, `readOntologyIri(Model)`, `readVersion(Model)`, `readVersion(Resource, Model)`, `readVersionDateTime(Model)`, `readVersionDateTime(Resource, Model)` and `readVersionIri(Model)` for metadata extraction
 * add **ExecutionRestController#getMetadata**: return metadata of loaded sources used in this execustion
 * add **UsePresentMappingProcessor**: Provides mappings for resources connected in the ontologies with given property paths.
+* add **TransitiveMappingProcessor**: Provides transitive closure of existing mappings.
 
 ### Fixed
 * fix **JaroWinklerMappingProcessor**: ignore other categories